Create comopnent for high score result and refactor - Tucker
