+++ components/HighScoreTableComponent.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { StyleSheet, View } from 'react-native';
 import { Text, useTheme } from 'react-native-paper';
 
@@ -28,8 +28,16 @@ export default function HighScoreTableComponent({ highScores, gameType, thisScor
       color: theme.colors.secondary,
       textAlign: 'right',
     },
+    nameText: {
+      minWidth: 50,
+      margin: 5,
+      fontSize: 20,
+      fontWeight: 'bold',
+      color: theme.colors.secondary,
+      textAlign: 'left',
+    },
     thisPlaceText: {
-      fontSize: 65,
+      fontSize: 30,
       color: theme.colors.primary,
       margin: 0
     },
@@ -55,6 +63,7 @@ export default function HighScoreTableComponent({ highScores, gameType, thisScor
       { highScores.map((highScore, index) => (
         <View key={highScore.id} style={styles.highScoreContainer}>
           <Text style={[styles.placeText, thisScore && thisScore.id === highScore.id ? styles.thisPlaceText : '']}>{index + 1}.</Text>
+          <Text style={[styles.nameText, thisScore && thisScore.id === highScore.id ? styles.thisPlaceText : '']}>{highScore.name}</Text>
           <Text style={[styles.highScoreText, thisScore && thisScore.id === highScore.id ? styles.thisPlaceText : '']}>{highScore.score}</Text>
         </View>
       ))}
